# openapi/spec.openapi.yaml
openapi: 3.0.2
servers:
  - description: OBADA Production API
    url: http://obs.node.obada.io
  - description: OBADA Staging API
    url: https://dev.node.obada.io
info:
  version: "0.0.1"
  title: OBADA API
  termsOfService: 'https://www.obada.io/terms/'
  contact:
    name: Obada Dev Team
    email: techops@obada.io
    url: 'https://www.obada.io'
...
security:
  - bearerAuth: []
paths:
  /accounts/my-balance:
    get:
      summary: Shows account balance of OBADA address
      operationId: balance
      tags:
        - Accounts
      responses:
        "200":
          $ref: "#/components/responses/AccountBalance"
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
           $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    AccountBalance:
      description: "Account balance response"
      content:
        application/json:
          schema:
            $ref: "definitions/Account.yml#/AccountBalance"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "Errors.yml#/InternalServerError"
...
    NotAuthorized:
      description: The request is not authorized.
      content:
        application/json:
          schema:
            $ref: "Errors.yml#/NotAuthorized"
    UnprocessableEntity:
      description: The submitted entity could not be processed.
      content:
        application/json:
          schema:
            $ref: "Errors.yml#/UnprocessableEntity"